backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
fredapi==0.5.2
apscheduler==3.10.4
pandas==2.2.3
numpy==2.2.1
pyarrow>=14.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
hmmlearn>=0.3.2
joblib>=1.3.0
PyYAML>=6.0.0
redis==5.2.1
httpx==0.28.1
python-dotenv==1.0.1
yfinance>=1.2.0
curl_cffi>=0.7
anthropic>=0.40.0
openai>=1.60.0
telethon>=1.36.0
